requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
dnspython==2.4.2
exceptiongroup==1.2.0
fonttools==4.47.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
loguru==0.7.2
matplotlib==3.8.2
multidict==6.0.4
numpy==1.26.2
openai==0.28.1
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pinecone-client==2.2.4
plotly==5.18.0
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
yarl==1.9.4
zipp==3.17.0